+++ inst/logout/app.R
@@ -5,4 +5,4 @@ library(auth0)
 ui <- fluidPage(logoutButton())
 
 server <- function(input, output, session) { }
-auth0App(ui, server)
+shinyAppAuth0(ui, server)